Some restructuring
