+++ app/dashboard/components/course-creation/lms-hover-content.tsx
